verbesserung

+++ config/createSettings.js
@@ -139,26 +139,27 @@ async function configure() {
 
 }
 
-if (process.argv[2] === "--override") {
-    fs.unlinkSync(file);
-    configure();
+if (require.main === module) {
+    if (process.argv[2] === "--override") {
+        fs.unlinkSync(file);
+        configure();
+    }
+
+    if (!fs.existsSync(file)) {
+        term.inverse('Es wurde kein config file gefunden! Konfiguration starten? [J|n]\n');
+
+        term.yesOrNo({yes: ['j', 'ENTER', 'J'], no: ['n', 'N']}, function (error, result) {
+
+            if (result) {
+                term("Die Konfiguration wird gestartet\n");
+                configure();
+            } else {
+                term("Die Konfiguration wird übersprungen.\n");
+                term.red("Denken Sie daran, die config Datei wie im README beschrieben zu erstellen, da Ella sonst nicht funktionieren wird!\n");
+                process.exit(0);
+            }
+        });
+    }
 }
-
-if (!fs.existsSync(file)) {
-    term.inverse('Es wurde kein config file gefunden! Konfiguration starten? [J|n]\n');
-
-    term.yesOrNo({yes: ['j', 'ENTER', 'J'], no: ['n', 'N']}, function (error, result) {
-
-        if (result) {
-            term("Die Konfiguration wird gestartet\n");
-            configure();
-        } else {
-            term("Die Konfiguration wird übersprungen.\n");
-            term.red("Denken Sie daran, die config Datei wie im README beschrieben zu erstellen, da Ella sonst nicht funktionieren wird!\n");
-            process.exit(0);
-        }
-    });
-}
-
 
 module.exports = {selectReleaseAndInstall};
